+++ src/components/checkbox/checkbox.test.ts
@@ -52,27 +52,42 @@ describe('<sl-checkbox>', () => {
     expect(el.invalid).to.be.false;
   });
 
-  it('should fire sl-change when clicked', async () => {
+  it('should emit sl-change and sl-input when clicked', async () => {
     const el = await fixture<SlCheckbox>(html` <sl-checkbox></sl-checkbox> `);
-    setTimeout(() => el.shadowRoot!.querySelector('input')!.click());
-    const event = (await oneEvent(el, 'sl-change')) as CustomEvent;
-    expect(event.target).to.equal(el);
+    const changeHandler = sinon.spy();
+    const inputHandler = sinon.spy();
+
+    el.addEventListener('sl-change', changeHandler);
+    el.addEventListener('sl-input', inputHandler);
+    el.click();
+    await el.updateComplete;
+
+    expect(changeHandler).to.have.been.calledOnce;
+    expect(inputHandler).to.have.been.calledOnce;
     expect(el.checked).to.be.true;
   });
 
-  it('should fire sl-change when toggled via keyboard', async () => {
+  it('should emit sl-change and sl-input when toggled with spacebar', async () => {
     const el = await fixture<SlCheckbox>(html` <sl-checkbox></sl-checkbox> `);
-    const input = el.shadowRoot!.querySelector('input')!;
-    input.focus();
-    setTimeout(() => sendKeys({ press: ' ' }));
-    const event = (await oneEvent(el, 'sl-change')) as CustomEvent;
-    expect(event.target).to.equal(el);
+    const changeHandler = sinon.spy();
+    const inputHandler = sinon.spy();
+
+    el.addEventListener('sl-change', changeHandler);
+    el.addEventListener('sl-input', inputHandler);
+    el.focus();
+    await el.updateComplete;
+    await sendKeys({ press: ' ' });
+
+    expect(changeHandler).to.have.been.calledOnce;
+    expect(inputHandler).to.have.been.calledOnce;
     expect(el.checked).to.be.true;
   });
 
-  it('should not fire sl-change when checked is set by javascript', async () => {
+  it('should not emit sl-change or sl-input when checked programmatically', async () => {
     const el = await fixture<SlCheckbox>(html` <sl-checkbox></sl-checkbox> `);
-    el.addEventListener('sl-change', () => expect.fail('event fired'));
+
+    el.addEventListener('sl-change', () => expect.fail('sl-change should not be emitted'));
+    el.addEventListener('sl-input', () => expect.fail('sl-input should not be emitted'));
     el.checked = true;
     await el.updateComplete;
     el.checked = false;