+++ src/components/rating/rating.test.ts
@@ -1,4 +1,7 @@
-import { expect, fixture, html } from '@open-wc/testing';
+import { expect, fixture, html, oneEvent } from '@open-wc/testing';
+import { sendKeys } from '@web/test-runner-commands';
+import sinon from 'sinon';
+import { clickOnElement, tapOnElement } from '../../internal/test';
 import type SlRating from './rating';
 
 describe('<sl-rating>', () => {
@@ -46,6 +49,41 @@ describe('<sl-rating>', () => {
     const base = el.shadowRoot!.querySelector<HTMLElement>('[part~="base"]')!;
 
     expect(base.getAttribute('aria-valuenow')).to.equal('3');
+  });
+
+  it('should emit sl-change when clicked', async () => {
+    const el = await fixture<SlRating>(html` <sl-rating></sl-rating> `);
+    const lastSymbol = el.shadowRoot!.querySelector<HTMLSpanElement>('.rating__symbol:last-child')!;
+    const changeHandler = sinon.spy();
+
+    el.addEventListener('sl-change', changeHandler);
+
+    await clickOnElement(lastSymbol);
+    await el.updateComplete;
+
+    expect(changeHandler).to.have.been.calledOnce;
+    expect(el.value).to.equal(5);
+  });
+
+  it('should emit sl-change when the value is changed with the keyboard', async () => {
+    const el = await fixture<SlRating>(html` <sl-rating></sl-rating> `);
+    const changeHandler = sinon.spy();
+
+    el.addEventListener('sl-change', changeHandler);
+    el.focus();
+    await el.updateComplete;
+    await sendKeys({ press: 'ArrowRight' });
+    await el.updateComplete;
+
+    expect(changeHandler).to.have.been.calledOnce;
+    expect(el.value).to.equal(1);
+  });
+
+  it('should not emit sl-change when the value is changed programmatically', async () => {
+    const el = await fixture<SlRating>(html` <sl-rating label="Test" value="1"></sl-rating> `);
+    el.addEventListener('sl-change', () => expect.fail('sl-change incorrectly emitted'));
+    el.value = 5;
+    await el.updateComplete;
   });
 
   describe('focus', () => {