+++ src/components/switch/switch.test.ts
@@ -1,5 +1,6 @@
 import { expect, fixture, html, oneEvent } from '@open-wc/testing';
 import { sendKeys } from '@web/test-runner-commands';
+import sinon from 'sinon';
 import type SlSwitch from './switch';
 
 describe('<sl-switch>', () => {
@@ -40,44 +41,72 @@ describe('<sl-switch>', () => {
     expect(el.invalid).to.be.false;
   });
 
-  it('should fire sl-change when clicked', async () => {
+  it('should emit sl-change and sl-input when clicked', async () => {
     const el = await fixture<SlSwitch>(html` <sl-switch></sl-switch> `);
-    setTimeout(() => el.shadowRoot!.querySelector('input')!.click());
-    const event = (await oneEvent(el, 'sl-change')) as CustomEvent;
-    expect(event.target).to.equal(el);
+    const changeHandler = sinon.spy();
+    const inputHandler = sinon.spy();
+
+    el.addEventListener('sl-change', changeHandler);
+    el.addEventListener('sl-input', inputHandler);
+    el.click();
+    await el.updateComplete;
+
+    expect(changeHandler).to.have.been.calledOnce;
+    expect(inputHandler).to.have.been.calledOnce;
     expect(el.checked).to.be.true;
   });
 
-  it('should fire sl-change when toggled with spacebar', async () => {
+  it('should emit sl-change when toggled with spacebar', async () => {
     const el = await fixture<SlSwitch>(html` <sl-switch></sl-switch> `);
+    const changeHandler = sinon.spy();
+    const inputHandler = sinon.spy();
+
+    el.addEventListener('sl-change', changeHandler);
+    el.addEventListener('sl-input', inputHandler);
     el.focus();
-    setTimeout(() => sendKeys({ press: ' ' }));
-    const event = (await oneEvent(el, 'sl-change')) as CustomEvent;
-    expect(event.target).to.equal(el);
+    await sendKeys({ press: ' ' });
+
+    expect(changeHandler).to.have.been.calledOnce;
+    expect(inputHandler).to.have.been.calledOnce;
     expect(el.checked).to.be.true;
   });
 
-  it('should fire sl-change when toggled with the right arrow', async () => {
+  it('should emit sl-change and sl-input when toggled with the right arrow', async () => {
     const el = await fixture<SlSwitch>(html` <sl-switch></sl-switch> `);
+    const changeHandler = sinon.spy();
+    const inputHandler = sinon.spy();
+
+    el.addEventListener('sl-change', changeHandler);
+    el.addEventListener('sl-input', inputHandler);
     el.focus();
-    setTimeout(() => sendKeys({ press: 'ArrowRight' }));
-    const event = (await oneEvent(el, 'sl-change')) as CustomEvent;
-    expect(event.target).to.equal(el);
+    await sendKeys({ press: 'ArrowRight' });
+    await el.updateComplete;
+
+    expect(changeHandler).to.have.been.calledOnce;
+    expect(inputHandler).to.have.been.calledOnce;
     expect(el.checked).to.be.true;
   });
 
-  it('should fire sl-change when toggled with the left arrow', async () => {
+  it('should emit sl-change and sl-input when toggled with the left arrow', async () => {
     const el = await fixture<SlSwitch>(html` <sl-switch checked></sl-switch> `);
+    const changeHandler = sinon.spy();
+    const inputHandler = sinon.spy();
+
+    el.addEventListener('sl-change', changeHandler);
+    el.addEventListener('sl-input', inputHandler);
     el.focus();
-    setTimeout(() => sendKeys({ press: 'ArrowLeft' }));
-    const event = (await oneEvent(el, 'sl-change')) as CustomEvent;
-    expect(event.target).to.equal(el);
+    await sendKeys({ press: 'ArrowLeft' });
+    await el.updateComplete;
+
+    expect(changeHandler).to.have.been.calledOnce;
+    expect(inputHandler).to.have.been.calledOnce;
     expect(el.checked).to.be.false;
   });
 
-  it('should not fire sl-change when checked is set by javascript', async () => {
+  it('should not emit sl-change or sl-input when checked is set by javascript', async () => {
     const el = await fixture<SlSwitch>(html` <sl-switch></sl-switch> `);
-    el.addEventListener('sl-change', () => expect.fail('event fired'));
+    el.addEventListener('sl-change', () => expect.fail('sl-change incorrectly emitted'));
+    el.addEventListener('sl-input', () => expect.fail('sl-change incorrectly emitted'));
     el.checked = true;
     await el.updateComplete;
     el.checked = false;